Refactor #188 - move package common buildscripts function
