+++ build-logic/convention/src/main/kotlin/com/nlab/reminder/Variants.kt
@@ -14,26 +14,13 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.convention
+package com.nlab.reminder
 
-import org.gradle.api.Project
-import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.kotlin.dsl.configure
-import org.gradle.kotlin.dsl.getByType
-import org.gradle.testing.jacoco.plugins.JacocoPluginExtension
+import com.android.build.api.variant.Variant
+import org.gradle.configurationcache.extensions.capitalized
 
 /**
  * @author Doohyun
  */
-internal fun Project.configureJacocoToolVersion() {
-    val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
-    configure<JacocoPluginExtension> {
-        toolVersion = libs.findVersion("jacoco").get().toString()
-    }
-}
-
-internal val jacocoExcludePatterns: Set<String>
-    get() = setOf(
-        "**/kotlin/**",
-        "**/infra/**",
-    )
+internal fun Variant.unitTestTaskName(): String? = unitTest?.let { "test${it.name.capitalized()}" }
+internal fun Variant.unitTestCapitalized(): String? = unitTest?.name?.capitalized()
